* commit 'a352c29bdc13a839cebd985b0de3ac1f13f09b91':
  combustionEngine: removed check for 'above full-load and below drag-load'
  removed duplicate property (drivingBehavior vs. driverBehavior), moved property from strategy to driver, moved vehiclestopped property to vehicle
  update iteration count inside loop
  disable writing search sequence
  formatting, use tolerance for assertions
  adapting expected values due to change of engine speed from int to double in AAUX
  gui: only log errors
  move constructors up
  added operating point to response, introduce abort criterion for search, coasting handles DistanceExceeded response (i.e. vehicle stops to early) and issues braking action
  code formatting
  code formatting
  [BusAux]: change engine speed from integer to double
  get correct cycle name and estimated driving time
  refactoring bus auxiliaries input provider
  use fulloadcurvereader

+++ VECTOAux/VectoAuxiliaries/Pneumatics/ICompressorMap.vb
@@ -26,23 +26,23 @@ Namespace Pneumatics
         ''' <param name="rpm">compressor rotation speed</param>
         ''' <returns></returns>
         ''' <remarks>Single</remarks>
-        Function GetFlowRate(ByVal rpm As Integer) As Single
+		Function GetFlowRate(ByVal rpm As Double) As Single
 
-        ''' <summary>
-        ''' Returns mechanical power at rpm when compressor is on
-        ''' </summary>
-        ''' <param name="rpm">compressor rotation speed</param>
-        ''' <returns></returns>
-        ''' <remarks>Single</remarks>
-        Function GetPowerCompressorOn(ByVal rpm As Integer) As Single
+		''' <summary>
+		''' Returns mechanical power at rpm when compressor is on
+		''' </summary>
+		''' <param name="rpm">compressor rotation speed</param>
+		''' <returns></returns>
+		''' <remarks>Single</remarks>
+		Function GetPowerCompressorOn(ByVal rpm As Double) As Single
 
-        ''' <summary>
-        ''' Returns mechanical power at rpm when compressor is off
-        ''' </summary>
-        ''' <param name="rpm">compressor rotation speed</param>
-        ''' <returns></returns>
-        ''' <remarks>Single</remarks>
-        Function GetPowerCompressorOff(ByVal rpm As Integer) As Single
+		''' <summary>
+		''' Returns mechanical power at rpm when compressor is off
+		''' </summary>
+		''' <param name="rpm">compressor rotation speed</param>
+		''' <returns></returns>
+		''' <remarks>Single</remarks>
+		Function GetPowerCompressorOff(ByVal rpm As Double) As Single
 
         'Returns Average Power Demand Per Compressor Unit FlowRate
         Function GetAveragePowerDemandPerCompressorUnitFlowRate() As Single
